requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
faiss-cpu
openai
pypdf2
dotenv
torch
sentence-transformers
transformers
langchain
langchain-huggingface
ollama